requirements.txt
streamlit
torch
torchvision
transformers
faiss-cpu
openai-whisper
opencv-python
Pillow
